Add files via upload
Jackpot number guessing game is the first personal project that I made from scratch. It uses random library to  generate a random number which corresponds to the jackpot number. This is the first time I learned the concept of loops. To add a nice touch in the end, I also used the counter to print the number to attempts it took to guess the jackpot number.